Use the constant expression instead of the result
A source of hard to detect errors is the use of the result of a constant expression claiming it represents some value. It is difficult to see if the constant is right. You can avoid this if you do not calculate it yourself, but let the parser/compiler/one time execution optimize the expression into the constant.

It also documents how you came up with the magic value and makes it easy to modify should you find you need some other constant.

There is #283 on this same constant. Not sure if you'd like to keep the underscores.